+++ cmd/option.go
@@ -30,6 +30,7 @@ type Options struct {
 	MinimumFailureSeverity string   `long:"minimum-failure-severity" description:"Sets minimum severity level for exiting with a non-zero error code" choice:"error" choice:"warning" choice:"notice"`
 	Color                  bool     `long:"color" description:"Enable colorized output"`
 	NoColor                bool     `long:"no-color" description:"Disable colorized output"`
+	Fix                    bool     `long:"fix" description:"Fix issues automatically"`
 	ActAsBundledPlugin     bool     `long:"act-as-bundled-plugin" hidden:"true"`
 }
 
@@ -83,30 +84,18 @@ func (opts *Options) toConfig() *tflint.Config {
 	}
 
 	rules := map[string]*tflint.RuleConfig{}
-	if len(opts.Only) > 0 {
-		// tflint-plugin-sdk v0.13+ doesn't need rules config when enabling the only option.
-		// This is for the backward compatibility.
-		for _, rule := range opts.Only {
-			rules[rule] = &tflint.RuleConfig{
-				Name:    rule,
-				Enabled: true,
-				Body:    nil,
-			}
+	for _, rule := range append(opts.Only, opts.EnableRules...) {
+		rules[rule] = &tflint.RuleConfig{
+			Name:    rule,
+			Enabled: true,
+			Body:    nil,
 		}
-	} else {
-		for _, rule := range opts.EnableRules {
-			rules[rule] = &tflint.RuleConfig{
-				Name:    rule,
-				Enabled: true,
-				Body:    nil,
-			}
-		}
-		for _, rule := range opts.DisableRules {
-			rules[rule] = &tflint.RuleConfig{
-				Name:    rule,
-				Enabled: false,
-				Body:    nil,
-			}
+	}
+	for _, rule := range opts.DisableRules {
+		rules[rule] = &tflint.RuleConfig{
+			Name:    rule,
+			Enabled: false,
+			Body:    nil,
 		}
 	}