+++ formatter/sarif_test.go
@@ -4,8 +4,10 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
+	"path/filepath"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
 	hcl "github.com/hashicorp/hcl/v2"
 	"github.com/terraform-linters/tflint/tflint"
 	"github.com/xeipuuv/gojsonschema"
@@ -21,7 +23,7 @@ func Test_sarifPrint(t *testing.T) {
 		{
 			Name:   "no issues",
 			Issues: tflint.Issues{},
-			Stdout: `{
+			Stdout: fmt.Sprintf(`{
   "version": "2.1.0",
   "$schema": "https://json.schemastore.org/sarif-2.1.0-rtm.5.json",
   "runs": [
@@ -29,6 +31,7 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint"
         }
       },
@@ -38,13 +41,14 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint-errors",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint"
         }
       },
       "results": []
     }
   ]
-}`,
+}`, tflint.Version, tflint.Version),
 		},
 		{
 			Name: "issues",
@@ -59,7 +63,7 @@ func Test_sarifPrint(t *testing.T) {
 					},
 				},
 			},
-			Stdout: `{
+			Stdout: fmt.Sprintf(`{
   "version": "2.1.0",
   "$schema": "https://json.schemastore.org/sarif-2.1.0-rtm.5.json",
   "runs": [
@@ -67,6 +71,7 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint",
           "rules": [
             {
@@ -108,29 +113,29 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint-errors",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint"
         }
       },
       "results": []
     }
   ]
-}`,
+}`, tflint.Version, tflint.Version),
 		},
 		{
-			Name: "Issues with SARIF-invalid position are output correctly",
+			Name: "issues not on line 1",
 			Issues: tflint.Issues{
 				{
 					Rule:    &testRule{},
 					Message: "test",
 					Range: hcl.Range{
 						Filename: "test.tf",
-						Start:    hcl.Pos{Line: 1, Column: 1},
-						End:      hcl.Pos{Line: 0, Column: 0},
+						Start:    hcl.Pos{Line: 3, Column: 1, Byte: 0},
+						End:      hcl.Pos{Line: 3, Column: 4, Byte: 3},
 					},
 				},
 			},
-			Error: fmt.Errorf("Failed to work; %w", errors.New("I don't feel like working")),
-			Stdout: `{
+			Stdout: fmt.Sprintf(`{
   "version": "2.1.0",
   "$schema": "https://json.schemastore.org/sarif-2.1.0-rtm.5.json",
   "runs": [
@@ -138,6 +143,79 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint",
+          "version": "%s",
+          "informationUri": "https://github.com/terraform-linters/tflint",
+          "rules": [
+            {
+              "id": "test_rule",
+              "shortDescription": {
+                "text": ""
+              },
+              "helpUri": "https://github.com"
+            }
+          ]
+        }
+      },
+      "results": [
+        {
+          "ruleId": "test_rule",
+          "level": "error",
+          "message": {
+            "text": "test"
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "artifactLocation": {
+                  "uri": "test.tf"
+                },
+                "region": {
+                  "startLine": 3,
+                  "startColumn": 1,
+                  "endLine": 3,
+                  "endColumn": 4
+                }
+              }
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "tool": {
+        "driver": {
+          "name": "tflint-errors",
+          "version": "%s",
+          "informationUri": "https://github.com/terraform-linters/tflint"
+        }
+      },
+      "results": []
+    }
+  ]
+}`, tflint.Version, tflint.Version),
+		},
+		{
+			Name: "issues spanning multiple lines",
+			Issues: tflint.Issues{
+				{
+					Rule:    &testRule{},
+					Message: "test",
+					Range: hcl.Range{
+						Filename: "test.tf",
+						Start:    hcl.Pos{Line: 1, Column: 1, Byte: 0},
+						End:      hcl.Pos{Line: 4, Column: 1, Byte: 3},
+					},
+				},
+			},
+			Stdout: fmt.Sprintf(`{
+  "version": "2.1.0",
+  "$schema": "https://json.schemastore.org/sarif-2.1.0-rtm.5.json",
+  "runs": [
+    {
+      "tool": {
+        "driver": {
+          "name": "tflint",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint",
           "rules": [
             {
@@ -166,7 +244,7 @@ func Test_sarifPrint(t *testing.T) {
                 "region": {
                   "startLine": 1,
                   "startColumn": 1,
-                  "endLine": 1,
+                  "endLine": 4,
                   "endColumn": 1
                 }
               }
@@ -179,6 +257,144 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint-errors",
+          "version": "%s",
+          "informationUri": "https://github.com/terraform-linters/tflint"
+        }
+      },
+      "results": []
+    }
+  ]
+}`, tflint.Version, tflint.Version),
+		},
+		{
+			Name: "issues in directories",
+			Issues: tflint.Issues{
+				{
+					Rule:    &testRule{},
+					Message: "test",
+					Range: hcl.Range{
+						Filename: filepath.Join("test", "main.tf"),
+						Start:    hcl.Pos{Line: 1, Column: 1, Byte: 0},
+						End:      hcl.Pos{Line: 1, Column: 4, Byte: 3},
+					},
+				},
+			},
+			Stdout: fmt.Sprintf(`{
+  "version": "2.1.0",
+  "$schema": "https://json.schemastore.org/sarif-2.1.0-rtm.5.json",
+  "runs": [
+    {
+      "tool": {
+        "driver": {
+          "name": "tflint",
+          "version": "%s",
+          "informationUri": "https://github.com/terraform-linters/tflint",
+          "rules": [
+            {
+              "id": "test_rule",
+              "shortDescription": {
+                "text": ""
+              },
+              "helpUri": "https://github.com"
+            }
+          ]
+        }
+      },
+      "results": [
+        {
+          "ruleId": "test_rule",
+          "level": "error",
+          "message": {
+            "text": "test"
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "artifactLocation": {
+                  "uri": "test/main.tf"
+                },
+                "region": {
+                  "startLine": 1,
+                  "startColumn": 1,
+                  "endLine": 1,
+                  "endColumn": 4
+                }
+              }
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "tool": {
+        "driver": {
+          "name": "tflint-errors",
+          "version": "%s",
+          "informationUri": "https://github.com/terraform-linters/tflint"
+        }
+      },
+      "results": []
+    }
+  ]
+}`, tflint.Version, tflint.Version),
+		},
+		{
+			Name: "Issues with missing source positions",
+			Issues: tflint.Issues{
+				{
+					Rule:    &testRule{},
+					Message: "test",
+					Range: hcl.Range{
+						Filename: "test.tf",
+					},
+				},
+			},
+			Error: fmt.Errorf("Failed to work; %w", errors.New("I don't feel like working")),
+			Stdout: fmt.Sprintf(`{
+  "version": "2.1.0",
+  "$schema": "https://json.schemastore.org/sarif-2.1.0-rtm.5.json",
+  "runs": [
+    {
+      "tool": {
+        "driver": {
+          "name": "tflint",
+          "version": "%s",
+          "informationUri": "https://github.com/terraform-linters/tflint",
+          "rules": [
+            {
+              "id": "test_rule",
+              "shortDescription": {
+                "text": ""
+              },
+              "helpUri": "https://github.com"
+            }
+          ]
+        }
+      },
+      "results": [
+        {
+          "ruleId": "test_rule",
+          "level": "error",
+          "message": {
+            "text": "test"
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "artifactLocation": {
+                  "uri": "test.tf"
+                }
+              }
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "tool": {
+        "driver": {
+          "name": "tflint-errors",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint"
         }
       },
@@ -193,7 +409,7 @@ func Test_sarifPrint(t *testing.T) {
       ]
     }
   ]
-}`,
+}`, tflint.Version, tflint.Version),
 		},
 		{
 			Name: "HCL diagnostics are surfaced as tflint-errors",
@@ -212,7 +428,7 @@ func Test_sarifPrint(t *testing.T) {
 					},
 				},
 			),
-			Stdout: `{
+			Stdout: fmt.Sprintf(`{
   "version": "2.1.0",
   "$schema": "https://json.schemastore.org/sarif-2.1.0-rtm.5.json",
   "runs": [
@@ -220,6 +436,7 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint"
         }
       },
@@ -229,6 +446,7 @@ func Test_sarifPrint(t *testing.T) {
       "tool": {
         "driver": {
           "name": "tflint-errors",
+          "version": "%s",
           "informationUri": "https://github.com/terraform-linters/tflint"
         }
       },
@@ -260,7 +478,7 @@ func Test_sarifPrint(t *testing.T) {
       ]
     }
   ]
-}`,
+}`, tflint.Version, tflint.Version),
 		},
 	}
 
@@ -272,8 +490,8 @@ func Test_sarifPrint(t *testing.T) {
 
 			formatter.Print(tc.Issues, tc.Error, map[string][]byte{})
 
-			if stdout.String() != tc.Stdout {
-				t.Fatalf("Failed %s test: expected=%s, stdout=%s", tc.Name, tc.Stdout, stdout.String())
+			if diff := cmp.Diff(tc.Stdout, stdout.String()); diff != "" {
+				t.Fatalf("Failed %s test: %s", tc.Name, diff)
 			}
 
 			schemaLoader := gojsonschema.NewReferenceLoader("http://json.schemastore.org/sarif-2.1.0")