+++ langserver/handler.go
@@ -11,6 +11,7 @@ import (
 	"runtime"
 	"strings"
 
+	"github.com/hashicorp/go-version"
 	"github.com/hashicorp/hcl/v2"
 	lsp "github.com/sourcegraph/go-lsp"
 	"github.com/sourcegraph/jsonrpc2"
@@ -43,12 +44,13 @@ func NewHandler(configPath string, cliConfig *tflint.Config) (jsonrpc2.Handler, 
 	}
 
 	rulesets := []tflint.RuleSet{}
+	clientSDKVersions := map[string]*version.Version{}
 	for name, ruleset := range rulsetPlugin.RuleSets {
 		constraints, err := ruleset.VersionConstraints()
 		if err != nil {
 			if st, ok := status.FromError(err); ok && st.Code() == codes.Unimplemented {
 				// VersionConstraints endpoint is available in tflint-plugin-sdk v0.14+.
-				// Skip verification if not available.
+				return nil, nil, fmt.Errorf(`Plugin "%s" SDK version is incompatible. Compatible versions: %s`, name, plugin.SDKVersionConstraints)
 			} else {
 				return nil, nil, fmt.Errorf("Failed to get TFLint version constraints to `%s` plugin; %w", name, err)
 			}
@@ -56,6 +58,20 @@ func NewHandler(configPath string, cliConfig *tflint.Config) (jsonrpc2.Handler, 
 		if !constraints.Check(tflint.Version) {
 			return nil, nil, fmt.Errorf("Failed to satisfy version constraints; tflint-ruleset-%s requires %s, but TFLint version is %s", name, constraints, tflint.Version)
 		}
+
+		clientSDKVersions[name], err = ruleset.SDKVersion()
+		if err != nil {
+			if st, ok := status.FromError(err); ok && st.Code() == codes.Unimplemented {
+				// SDKVersion endpoint is available in tflint-plugin-sdk v0.14+.
+				return nil, nil, fmt.Errorf(`Plugin "%s" SDK version is incompatible. Compatible versions: %s`, name, plugin.SDKVersionConstraints)
+			} else {
+				return nil, nil, fmt.Errorf(`Failed to get plugin "%s" SDK version; %w`, name, err)
+			}
+		}
+		if !plugin.SDKVersionConstraints.Check(clientSDKVersions[name]) {
+			return nil, nil, fmt.Errorf(`Plugin "%s" SDK version (%s) is incompatible. Compatible versions: %s`, name, clientSDKVersions[name], plugin.SDKVersionConstraints)
+		}
+
 		rulesets = append(rulesets, ruleset)
 	}
 	if err := cliConfig.ValidateRules(rulesets...); err != nil {
@@ -63,24 +79,26 @@ func NewHandler(configPath string, cliConfig *tflint.Config) (jsonrpc2.Handler, 
 	}
 
 	return jsonrpc2.HandlerWithError((&handler{
-		configPath: configPath,
-		cliConfig:  cliConfig,
-		config:     cfg,
-		fs:         afero.NewCopyOnWriteFs(afero.NewOsFs(), afero.NewMemMapFs()),
-		plugin:     rulsetPlugin,
-		diagsPaths: []string{},
+		configPath:        configPath,
+		cliConfig:         cliConfig,
+		config:            cfg,
+		fs:                afero.NewCopyOnWriteFs(afero.NewOsFs(), afero.NewMemMapFs()),
+		plugin:            rulsetPlugin,
+		clientSDKVersions: clientSDKVersions,
+		diagsPaths:        []string{},
 	}).handle), rulsetPlugin, nil
 }
 
 type handler struct {
-	configPath string
-	cliConfig  *tflint.Config
-	config     *tflint.Config
-	fs         afero.Fs
-	rootDir    string
-	plugin     *plugin.Plugin
-	shutdown   bool
-	diagsPaths []string
+	configPath        string
+	cliConfig         *tflint.Config
+	config            *tflint.Config
+	fs                afero.Fs
+	rootDir           string
+	plugin            *plugin.Plugin
+	clientSDKVersions map[string]*version.Version
+	shutdown          bool
+	diagsPaths        []string
 }
 
 func (h *handler) handle(ctx context.Context, conn *jsonrpc2.Conn, req *jsonrpc2.Request) (result interface{}, err error) {
@@ -210,7 +228,7 @@ func (h *handler) inspect() (map[string][]lsp.Diagnostic, error) {
 			return ret, fmt.Errorf("Failed to apply config to `%s` plugin", name)
 		}
 		for _, runner := range runners {
-			err = ruleset.Check(plugin.NewGRPCServer(runner, runners[len(runners)-1], loader.Files()))
+			err = ruleset.Check(plugin.NewGRPCServer(runner, runners[len(runners)-1], loader.Files(), h.clientSDKVersions[name]))
 			if err != nil {
 				return ret, fmt.Errorf("Failed to check ruleset: %w", err)
 			}